+++ frontend/src/components/Tab.jsx
@@ -2,7 +2,7 @@
 function Tab(props) {
     return (
         <span>
-            <button onClick={() => {props.setChosenCategory(props.category)}} type="button" className={props.isChosen? "btn btn-dark rounded m-2" : "btn btn-outline-dark rounded m-2"}>{props.category}</button>
+            <button onClick={() => { props.setChosenCategory(props.category) }} type="button" className={props.isChosen ? "btn btn-dark rounded m-2" : "btn btn-outline-dark rounded m-2"}>{props.category}</button>
         </span>
     )
 }
